+++ firebase-login/public/register.html
@@ -16,7 +16,7 @@
             <hr>
             
             <label for="email"><b>Email</b></label>
-            <input type="text" placeholder="Enter Email" name="email" id="email" required>
+            <input type="text" autocomplete="off" placeholder="Enter Email" name="email" id="email" required>
     
             <label><b>Password</b></label>
             <input type="password" placeholder="Password" name="psw" id="psw" required>
@@ -33,64 +33,7 @@
     </form>
 
 
-    <script type="module">
-        import { initializeApp, } from "https://www.gstatic.com/firebasejs/9.13.0/firebase-app.js";
-        import { getAuth,createUserWithEmailAndPassword,signInWithEmailAndPassword,signOut,onAuthStateChanged} from "https://www.gstatic.com/firebasejs/9.13.0/firebase-auth.js";
-        import { getFirestore,collection,setDoc,doc } from "https://www.gstatic.com/firebasejs/9.13.0/firebase-firestore.js";
-
-        // TODO: Add SDKs for Firebase products that you want to use
-        // https://firebase.google.com/docs/web/setup#available-libraries
-
-        // Your web app's Firebase configuration
-        // For Firebase JS SDK v7.20.0 and later, measurementId is optional
-        const firebaseConfig = {
-            apiKey: "AIzaSyAUS-fQ46Gv4kuLEqvYurwQJ5inCRkLA1U",
-            authDomain: "plantyourdreams-8d08e.firebaseapp.com",
-            databaseURL: "https://plantyourdreams-8d08e-default-rtdb.firebaseio.com",
-            projectId: "plantyourdreams-8d08e",
-            storageBucket: "plantyourdreams-8d08e.appspot.com",
-            messagingSenderId: "63192119407",
-            appId: "1:63192119407:web:c15c2b763e51e7f786b59a",
-            measurementId: "G-XYRQCEW5L5"
-        };
-
-        // Initialize Firebase
-        const app = initializeApp(firebaseConfig);
-        const auth = getAuth(app);
-        const db = getFirestore(app);
-
-        registerbtn.addEventListener('click',(e) => {
-            var email = document.getElementById('email').value;
-            var password = document.getElementById('psw').value;
-
-
-            createUserWithEmailAndPassword(auth, email, password)
-            .then((userCredential) => {
-                // Signed in
-                const user = userCredential.user;
-                var lgDate = new Date();
-                alert("New User Created!")
-
-                setDoc(doc(db, 'users', (user.uid)), {
-                    email: user.email,
-                    lastLogin: lgDate 
-                });
-            })
-            .catch((error) => {
-                const errorCode = error.code;
-                const errorMessage = error.message;
-                // ..
-                alert(errorMessage);
-            });
-
-
-            // signOut(auth).then(() => {
-            //     // Sign-out successful.
-            // }).catch((error) => {
-            //     // An error happened.
-            // });
-        })
-    </script>
+    <script type="module" src="./register.js"></script>
     
 </body>
 </html>
